+++ grails-app/domain/com/caseritos/Estado.groovy
@@ -1,4 +1,4 @@
-package com.torneo.fifa
+package com.caseritos
 
 class Estado {
