+++ grails-app/taglib/com/caseritos/LoginTagLib.groovy
@@ -1,4 +1,4 @@
-package com.torneo.fifa
+package com.caseritos
 
 
 class LoginTagLib {
