+++ test/unit/com/caseritos/PartidoControllerSpec.groovy
@@ -1,13 +1,13 @@
-package com.torneo.fifa
+package com.caseritos
 
 
 
 import grails.test.mixin.*
 import spock.lang.*
 
-@TestFor(TorneoController)
-@Mock(Torneo)
-class TorneoControllerSpec extends Specification {
+@TestFor(PartidoController)
+@Mock(Partido)
+class PartidoControllerSpec extends Specification {
 
     def populateValidParams(params) {
         assert params != null
@@ -21,8 +21,8 @@ class TorneoControllerSpec extends Specification {
             controller.index()
 
         then:"The model is correct"
-            !model.torneoInstanceList
-            model.torneoInstanceCount == 0
+            !model.partidoInstanceList
+            model.partidoInstanceCount == 0
     }
 
     void "Test the create action returns the correct model"() {
@@ -30,7 +30,7 @@ class TorneoControllerSpec extends Specification {
             controller.create()
 
         then:"The model is correctly created"
-            model.torneoInstance!= null
+            model.partidoInstance!= null
     }
 
     void "Test the save action correctly persists an instance"() {
@@ -38,25 +38,25 @@ class TorneoControllerSpec extends Specification {
         when:"The save action is executed with an invalid instance"
             request.contentType = FORM_CONTENT_TYPE
             request.method = 'POST'
-            def torneo = new Torneo()
-            torneo.validate()
-            controller.save(torneo)
+            def partido = new Partido()
+            partido.validate()
+            controller.save(partido)
 
         then:"The create view is rendered again with the correct model"
-            model.torneoInstance!= null
+            model.partidoInstance!= null
             view == 'create'
 
         when:"The save action is executed with a valid instance"
             response.reset()
             populateValidParams(params)
-            torneo = new Torneo(params)
+            partido = new Partido(params)
 
-            controller.save(torneo)
+            controller.save(partido)
 
         then:"A redirect is issued to the show action"
-            response.redirectedUrl == '/torneo/show/1'
+            response.redirectedUrl == '/partido/show/1'
             controller.flash.message != null
-            Torneo.count() == 1
+            Partido.count() == 1
     }
 
     void "Test that the show action returns the correct model"() {
@@ -68,11 +68,11 @@ class TorneoControllerSpec extends Specification {
 
         when:"A domain instance is passed to the show action"
             populateValidParams(params)
-            def torneo = new Torneo(params)
-            controller.show(torneo)
+            def partido = new Partido(params)
+            controller.show(partido)
 
         then:"A model is populated containing the domain instance"
-            model.torneoInstance == torneo
+            model.partidoInstance == partido
     }
 
     void "Test that the edit action returns the correct model"() {
@@ -84,11 +84,11 @@ class TorneoControllerSpec extends Specification {
 
         when:"A domain instance is passed to the edit action"
             populateValidParams(params)
-            def torneo = new Torneo(params)
-            controller.edit(torneo)
+            def partido = new Partido(params)
+            controller.edit(partido)
 
         then:"A model is populated containing the domain instance"
-            model.torneoInstance == torneo
+            model.partidoInstance == partido
     }
 
     void "Test the update action performs an update on a valid domain instance"() {
@@ -98,28 +98,28 @@ class TorneoControllerSpec extends Specification {
             controller.update(null)
 
         then:"A 404 error is returned"
-            response.redirectedUrl == '/torneo/index'
+            response.redirectedUrl == '/partido/index'
             flash.message != null
 
 
         when:"An invalid domain instance is passed to the update action"
             response.reset()
-            def torneo = new Torneo()
-            torneo.validate()
-            controller.update(torneo)
+            def partido = new Partido()
+            partido.validate()
+            controller.update(partido)
 
         then:"The edit view is rendered again with the invalid instance"
             view == 'edit'
-            model.torneoInstance == torneo
+            model.partidoInstance == partido
 
         when:"A valid domain instance is passed to the update action"
             response.reset()
             populateValidParams(params)
-            torneo = new Torneo(params).save(flush: true)
-            controller.update(torneo)
+            partido = new Partido(params).save(flush: true)
+            controller.update(partido)
 
         then:"A redirect is issues to the show action"
-            response.redirectedUrl == "/torneo/show/$torneo.id"
+            response.redirectedUrl == "/partido/show/$partido.id"
             flash.message != null
     }
 
@@ -130,23 +130,23 @@ class TorneoControllerSpec extends Specification {
             controller.delete(null)
 
         then:"A 404 is returned"
-            response.redirectedUrl == '/torneo/index'
+            response.redirectedUrl == '/partido/index'
             flash.message != null
 
         when:"A domain instance is created"
             response.reset()
             populateValidParams(params)
-            def torneo = new Torneo(params).save(flush: true)
+            def partido = new Partido(params).save(flush: true)
 
         then:"It exists"
-            Torneo.count() == 1
+            Partido.count() == 1
 
         when:"The domain instance is passed to the delete action"
-            controller.delete(torneo)
+            controller.delete(partido)
 
         then:"The instance is deleted"
-            Torneo.count() == 0
-            response.redirectedUrl == '/torneo/index'
+            Partido.count() == 0
+            response.redirectedUrl == '/partido/index'
             flash.message != null
     }
 }
