+++ test/unit/com/caseritos/PlayerSpec.groovy
@@ -1,4 +1,4 @@
-package com.torneo.fifa
+package com.caseritos
 
 import grails.test.mixin.TestFor
 import spock.lang.Specification
@@ -6,8 +6,8 @@ import spock.lang.Specification
 /**
  * See the API for {@link grails.test.mixin.domain.DomainClassUnitTestMixin} for usage instructions
  */
-@TestFor(Torneo)
-class TorneoSpec extends Specification {
+@TestFor(Player)
+class PlayerSpec extends Specification {
 
     def setup() {
     }
