+++ test/unit/com/caseritos/TorneoControllerSpec.groovy
@@ -1,13 +1,13 @@
-package com.torneo.fifa
+package com.caseritos
 
 
 
 import grails.test.mixin.*
 import spock.lang.*
 
-@TestFor(PartidoController)
-@Mock(Partido)
-class PartidoControllerSpec extends Specification {
+@TestFor(TorneoController)
+@Mock(Torneo)
+class TorneoControllerSpec extends Specification {
 
     def populateValidParams(params) {
         assert params != null
@@ -21,8 +21,8 @@ class PartidoControllerSpec extends Specification {
             controller.index()
 
         then:"The model is correct"
-            !model.partidoInstanceList
-            model.partidoInstanceCount == 0
+            !model.torneoInstanceList
+            model.torneoInstanceCount == 0
     }
 
     void "Test the create action returns the correct model"() {
@@ -30,7 +30,7 @@ class PartidoControllerSpec extends Specification {
             controller.create()
 
         then:"The model is correctly created"
-            model.partidoInstance!= null
+            model.torneoInstance!= null
     }
 
     void "Test the save action correctly persists an instance"() {
@@ -38,25 +38,25 @@ class PartidoControllerSpec extends Specification {
         when:"The save action is executed with an invalid instance"
             request.contentType = FORM_CONTENT_TYPE
             request.method = 'POST'
-            def partido = new Partido()
-            partido.validate()
-            controller.save(partido)
+            def torneo = new Torneo()
+            torneo.validate()
+            controller.save(torneo)
 
         then:"The create view is rendered again with the correct model"
-            model.partidoInstance!= null
+            model.torneoInstance!= null
             view == 'create'
 
         when:"The save action is executed with a valid instance"
             response.reset()
             populateValidParams(params)
-            partido = new Partido(params)
+            torneo = new Torneo(params)
 
-            controller.save(partido)
+            controller.save(torneo)
 
         then:"A redirect is issued to the show action"
-            response.redirectedUrl == '/partido/show/1'
+            response.redirectedUrl == '/torneo/show/1'
             controller.flash.message != null
-            Partido.count() == 1
+            Torneo.count() == 1
     }
 
     void "Test that the show action returns the correct model"() {
@@ -68,11 +68,11 @@ class PartidoControllerSpec extends Specification {
 
         when:"A domain instance is passed to the show action"
             populateValidParams(params)
-            def partido = new Partido(params)
-            controller.show(partido)
+            def torneo = new Torneo(params)
+            controller.show(torneo)
 
         then:"A model is populated containing the domain instance"
-            model.partidoInstance == partido
+            model.torneoInstance == torneo
     }
 
     void "Test that the edit action returns the correct model"() {
@@ -84,11 +84,11 @@ class PartidoControllerSpec extends Specification {
 
         when:"A domain instance is passed to the edit action"
             populateValidParams(params)
-            def partido = new Partido(params)
-            controller.edit(partido)
+            def torneo = new Torneo(params)
+            controller.edit(torneo)
 
         then:"A model is populated containing the domain instance"
-            model.partidoInstance == partido
+            model.torneoInstance == torneo
     }
 
     void "Test the update action performs an update on a valid domain instance"() {
@@ -98,28 +98,28 @@ class PartidoControllerSpec extends Specification {
             controller.update(null)
 
         then:"A 404 error is returned"
-            response.redirectedUrl == '/partido/index'
+            response.redirectedUrl == '/torneo/index'
             flash.message != null
 
 
         when:"An invalid domain instance is passed to the update action"
             response.reset()
-            def partido = new Partido()
-            partido.validate()
-            controller.update(partido)
+            def torneo = new Torneo()
+            torneo.validate()
+            controller.update(torneo)
 
         then:"The edit view is rendered again with the invalid instance"
             view == 'edit'
-            model.partidoInstance == partido
+            model.torneoInstance == torneo
 
         when:"A valid domain instance is passed to the update action"
             response.reset()
             populateValidParams(params)
-            partido = new Partido(params).save(flush: true)
-            controller.update(partido)
+            torneo = new Torneo(params).save(flush: true)
+            controller.update(torneo)
 
         then:"A redirect is issues to the show action"
-            response.redirectedUrl == "/partido/show/$partido.id"
+            response.redirectedUrl == "/torneo/show/$torneo.id"
             flash.message != null
     }
 
@@ -130,23 +130,23 @@ class PartidoControllerSpec extends Specification {
             controller.delete(null)
 
         then:"A 404 is returned"
-            response.redirectedUrl == '/partido/index'
+            response.redirectedUrl == '/torneo/index'
             flash.message != null
 
         when:"A domain instance is created"
             response.reset()
             populateValidParams(params)
-            def partido = new Partido(params).save(flush: true)
+            def torneo = new Torneo(params).save(flush: true)
 
         then:"It exists"
-            Partido.count() == 1
+            Torneo.count() == 1
 
         when:"The domain instance is passed to the delete action"
-            controller.delete(partido)
+            controller.delete(torneo)
 
         then:"The instance is deleted"
-            Partido.count() == 0
-            response.redirectedUrl == '/partido/index'
+            Torneo.count() == 0
+            response.redirectedUrl == '/torneo/index'
             flash.message != null
     }
 }
